chore: some refactor for socket
