+++ Backend/package.json
@@ -17,6 +17,7 @@
   "dependencies": {
     "@types/socket.io": "^3.0.2",
     "dotenv": "^16.3.1",
+    "mongoose": "^8.0.0",
     "socket.io": "^4.7.2",
     "ts-node": "^10.9.1"
   }
