+++ Backend/tsconfig.json
@@ -9,6 +9,6 @@
     "skipLibCheck": true,
     "forceConsistentCasingInFileNames": true
   },
-  "include": ["src/**/*.ts"],
+  "include": ["src/**/*.ts", "env.d.ts"],
   "exclude": ["node_modules"]
 }
